requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.1
requests==2.32.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
python-decouple==3.8
python-dotenv==1.0.0
dj-database-url==2.1.0
